# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Simva, AKA Simple Validator, AKA SurveyManager is the evolution of a tool created to integrate and ease the task of validating serious games using pre-post surveys as a control point. This version allows to create more complex studies with more variety of activities and different allocations.",
    "version": "1.0.0",
    "title": "Simva API",
    "contact": {
      "email": "ivanjper@ucm.es"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/studies/{studyid}/tests/{testid}/activities": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the list of activities for the selected test",
        "operationId": "getTestActivities",
        "description": "Obtains the list of activities for the selected test\n",
        "responses": {
          "200": {
            "description": "the list of activities for this test",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/activity"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a Activity to a test",
        "operationId": "addActivityToTest",
        "description": "Creates a new activity for the current teacher.\n",
        "responses": {
          "200": {
            "description": "activity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyStudyId"
          },
          {
            "$ref": "#/components/parameters/testId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activity"
              }
            }
          }
        }
      }
    },
...
    "/activities": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the list of activities for the current teacher",
        "operationId": "getActivities",
        "description": "Obtains the list of activities for the current teacher.\n",
        "responses": {
          "200": {
            "description": "the list of activities for this teacher",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/activity"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a Activity",
        "operationId": "addActivity",
        "description": "Creates a new activity for the current teacher.\n",
        "responses": {
          "200": {
            "description": "activity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activity"
              }
            }
          },
          "description": "Activity to be created"
        }
      }
    },
    "/activities/{id}": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the activity with the given ID",
        "operationId": "getActivity",
        "description": "Obtains the requested activity\n",
        "responses": {
          "200": {
            "description": "Activity object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      },
      "put": {
        "tags": [
          "teachers"
        ],
        "summary": "updates the test",
        "operationId": "updateActivity",
        "description": "Updates an existing test",
        "responses": {
          "200": {
            "description": "item updated"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Teacher does not own the class"
          },
          "404": {
            "description": "Requested study does not exist"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/studyId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activity"
              }
            }
          },
          "description": "activity to be modified"
        }
      },
      "patch": {
        "summary": "Modifies the activity.",
        "tags": [
          "teachers"
        ],
        "operationId": "patchActivity",
        "description": "Modifies the activity\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activityBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activity updated",
            "schema": {
              "$ref": "#/components/schemas/activity"
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "teachers"
        ],
        "summary": "deletes the activity with the given ID",
        "operationId": "deleteActivity",
        "description": "Deleted the specified activity\n",
        "responses": {
          "200": {
            "description": "activity deleted"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/export": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the activity export with the given ID",
        "operationId": "exportActivity",
        "description": "Obtains the requested activity export\n",
        "responses": {
          "200": {
            "description": "Activity export object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/surveyowner": {
      "patch": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "update the survey owner for a Limesurvey activity",
        "operationId": "updateSurveyOwner",
        "description": "updateSurveyOwner\n",
        "responses": {
          "200": {
            "description": "Survey Updated"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/usersurveylist": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "get user survey list for a Limesurvey activity",
        "operationId": "getUserSurveyList",
        "description": "getUserSurveyList\n",
        "responses": {
          "200": {
            "description": "Survey List"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/openable": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "returns if the activity can be opened or not",
        "operationId": "getOpenable",
        "description": "If the activity can be opened, e.g. is hosted in a web,\nthe activity will be openable as can be opened.\n",
        "responses": {
          "200": {
            "description": "if it is openable or not",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/open": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "redirects the user to the activity landing",
        "operationId": "openActivity",
        "description": "If the activity is openable, redirects the user to\nthe activity landing.\n",
        "responses": {
          "307": {
            "description": "redirects to the valid landing"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/target": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "returns if the activity can be opened or not",
        "operationId": "getActivityTarget",
        "description": "If the activity can be opened, e.g. is hosted in a web,\nthis endpoint will return the target URI\n",
        "responses": {
          "200": {
            "description": "the target URI of the activity"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          }
        ]
      }
    },
    "/activities/{id}/progress": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "gets the progress status of the activity",
        "operationId": "getCompletion",
        "description": "Obtains the progress status of the activity\n",
        "responses": {
          "200": {
            "description": "if it is openable or not"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          },
          {
            "$ref": "#/components/parameters/users"
          }
        ]
      }
    },
    "/activities/{id}/completion": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "gets the completion status of the activity",
        "operationId": "getCompletion",
        "description": "Obtains the completion status of the activity\n",
        "responses": {
          "200": {
            "description": "if it is openable or not"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          },
          {
            "$ref": "#/components/parameters/users"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "sets the completion status of the activity",
        "operationId": "setCompletion",
        "description": "Set the completion status of the activity for a student\n",
        "responses": {
          "200": {
            "description": "if it is openable or not"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          },
          {
            "$ref": "#/components/parameters/user"
          }
        ]
      }
    },
    "/activities/{id}/result": {
      "get": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "gets the results of the activity",
        "operationId": "getResult",
        "description": "Obtains the completion status of the activity\n",
        "responses": {
          "200": {
            "description": "the result of the students required"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          },
          {
            "$ref": "#/components/parameters/users"
          },
          {
            "$ref": "#/components/parameters/type"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers",
          "students"
        ],
        "summary": "sets the result for the activity",
        "operationId": "setResult",
        "description": "Set the result of the activity for a student\n",
        "responses": {
          "200": {
            "description": "if it is openable or not"
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/testId"
          },
          {
            "$ref": "#/components/parameters/user"
          }
        ]
      }
    },
...
  },
  "security": [
    {
      "OAuth2": [
        "read",
        "write"
      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/auth",
            "tokenUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/token",
            "scopes": {
              "read": "Grants read access",
              "write": "Grants write access"
            }
          }
        }
      }
    },
    "parameters": {
...
      "studyStudyId": {
        "in": "path",
        "name": "studyid",
        "required": true,
        "schema": {
          "type": "string",
          "example": "5d07a0767c1405004bc702af"
        },
        "description": "The study ID"
      },
      "testId": {
        "in": "path",
        "name": "testid",
        "required": true,
        "schema": {
          "type": "string",
          "example": "5d07a0767c1405004bc702at"
        },
        "description": "The test ID"
      },
      "searchString": {
        "in": "query",
        "name": "searchString",
        "description": "pass an optional search string for result filtering",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
  }
}